LayoutTests/inspector/styles/inject-stylesheet-expected.txt

CONSOLE MESSAGE: line 9: iframe loaded
Tests that injected user stylesheets are reflected in the Styles pane.


Main frame style:
[expanded] 

[expanded] 
element.style  { ()

======== Matched CSS Rules ========
[expanded] 
media="screen" (inject-stylesheet.html)
#main  { (inject-stylesheet.html:5)
background: blue;
    background-image: initial;
    background-position-x: initial;
    background-position-y: initial;
    background-size: initial;
    background-repeat-x: initial;
    background-repeat-y: initial;
    background-attachment: initial;
    background-origin: initial;
    background-clip: initial;
    background-color: blue;

[expanded] 
#main  { (user stylesheet)
color: red;

[expanded] 
div  { (user agent stylesheet)
display: block;


iframe style:
[expanded] 

[expanded] 
element.style  { ()

======== Matched CSS Rules ========
[expanded] 
media="screen" (inject-stylesheet-iframe-data.html)
body  { (inject-stylesheet-iframe-data.html:4)
background: green;
    background-image: initial;
    background-position-x: initial;
    background-position-y: initial;
    background-size: initial;
    background-repeat-x: initial;
    background-repeat-y: initial;
    background-attachment: initial;
    background-origin: initial;
    background-clip: initial;
    background-color: green;

[expanded] 
#iframebody  { (user stylesheet)
/-- overloaded --/ background: red;
    /-- overloaded --/ background-image: initial;
    /-- overloaded --/ background-position-x: initial;
    /-- overloaded --/ background-position-y: initial;
    /-- overloaded --/ background-size: initial;
    /-- overloaded --/ background-repeat-x: initial;
    /-- overloaded --/ background-repeat-y: initial;
    /-- overloaded --/ background-attachment: initial;
    /-- overloaded --/ background-origin: initial;
    /-- overloaded --/ background-clip: initial;
    /-- overloaded --/ background-color: red;

[expanded] 
body  { (user agent stylesheet)
display: block;
margin: 8px;
    margin-top: 8px;
    margin-right: 8px;
    margin-bottom: 8px;
    margin-left: 8px;
